Auth API HttpStatus 수정
