+++ src/main/java/com/sprint/mission/discodeit/dto/request/LoginRequestDto.java
@@ -1,10 +1,12 @@
 package com.sprint.mission.discodeit.dto.request;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
+@Schema(description = "로그인 정보")
 public class LoginRequestDto {
     private final String username;
     private final String password;
